+++ src/main/java/me/fallenbreath/yamlang/ConverterConfiguration.java
@@ -1,0 +1,10 @@
+package me.fallenbreath.yamlang;
+
+public class ConverterConfiguration
+{
+    public final boolean allowLists;
+    public ConverterConfiguration(boolean allowLists)
+    {
+        this.allowLists = allowLists;
+    }
+}
